requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Werkzeug==3.1.3
